ai/ieng_ai/requirements.txt
appdirs==1.4.4
asgiref==3.5.2
asttokens==2.0.8
audioread==3.0.0
backcall==0.2.0
boto3==1.24.82
botocore==1.27.82
certifi==2022.6.15
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.1
ci-sdr==0.0.2
clarifai==2.6.2
clarifai-grpc==8.8.0
click==8.1.3
colorama==0.4.5
ConfigArgParse==1.5.3
configparser==3.8.1
contourpy==1.0.5
ctc-segmentation==1.7.3
cycler==0.11.0
Cython==0.29.32
decorator==5.1.1
Distance==0.1.3
Django==4.1.1
django-extensions==3.2.1
django-storages==1.13.1
djangorestframework==3.14.0
einops==0.4.1
espnet==202207
espnet-model-zoo==0.1.7
espnet-tts-frontend==0.0.3
executing==1.1.0
fast-bss-eval==0.1.3
filelock==3.8.0
fonttools==4.37.4
future==0.18.2
g2p-en==2.1.0
googleapis-common-protos==1.53.0
grpcio==1.44.0
h5py==3.7.0
huggingface-hub==0.10.0
humanfriendly==10.0
idna==2.10
inflect==6.0.0
ipython==8.5.0
jaconv==0.3
jamo==0.4.1
jedi==0.18.1
jmespath==1.0.1
joblib==1.2.0
jsonschema==2.6.0
kaldiio==2.17.2
kiwisolver==1.4.4
librosa==0.9.2
llvmlite==0.39.1
matplotlib==3.6.0
matplotlib-inline==0.1.6
mysql-connector-python==8.0.30
nltk==3.7
numba==0.56.2
numpy==1.23.3
packaging==21.3
pandas==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
pooch==1.6.0
prompt-toolkit==3.0.31
protobuf==3.19.3
pure-eval==0.2.2
pycparser==2.21
pydantic==1.10.2
Pygments==2.13.0
pyparsing==3.0.9
pypinyin==0.44.0
pyreadline3==3.4.1
python-dateutil==2.8.2
pytorch-wpe==0.0.1
pytz==2022.2.1
pyworld==0.3.0
PyYAML==6.0
regex==2022.9.13
requests==2.25.1
resampy==0.4.2
s3transfer==0.6.0
scikit-learn==1.1.2
scipy==1.9.1
sentencepiece==0.1.97
six==1.16.0
soundfile==0.11.0
sqlparse==0.4.3
stack-data==0.5.1
threadpoolctl==3.1.0
torch==1.12.1
torch-complex==0.4.3
tqdm==4.64.1
traitlets==5.4.0
typeguard==2.13.3
typing_extensions==4.3.0
tzdata==2022.4
Unidecode==1.3.6
urllib3==1.26.12
wcwidth==0.2.5
